gsk/gl: use default_framebuffer in render job
Now that we have the ability to set a default framebuffer for a
GdkGLContext, the GskGLRenderer needs to respect that when creating the
render job.  Furthermore, we need to check against it when applying the
scissor test.

Generally, 0 is the default framebuffer. But on platforms where we need
to bind a platform-specific feature to a GL_FRAMEBUFFER, we might have a
default that is not 0. For example, on macOS we bind an IOSurfaceRef to
a GL_TEXTURE_RECTANGLE which then is assigned as the backing store for a
framebuffer we want as the destination. This is different than using
gsk_gl_renderer_render_texture() in that we don't want to incur an extra
copy to the destination surface.